rewrite test script in bash
The HTTP server is now started in a subprocess, not a thread. Running it
in a Ruby thread had all sorts of problems on different Rubies, such as
random failures on Rubinius.

The new script also avoids starting the HTTP server if not needed, e.g.
if filtering arguments have been used to select only the non-adapter tests.